\name{selectAbsences}
\alias{selectAbsences}
\title{
Select (spatially biased) absence rows.
}
\description{
This function takes a matrix or (optionally spatial) data frame containing species presence (1) and (pseudo)absence (0) data (e.g., the output of \code{\link{gridRecords}}), and it selects among the absence rows to stay within a given number or ratio of absences, and/or within and/or beyond a given distance to the presences. Optionally, to incorporate survey bias, absences can be selected with higher probability towards the vicinity of presences; or with probability weights according to a user-provided bias raster layer.
}
\usage{
selectAbsences(data, sp.cols, coord.cols = NULL, CRS = NULL, min.dist = NULL,
max.dist = NULL, n = NULL, mult.p = NULL, bias = FALSE, bunch = FALSE,
dist.mat = NULL, seed = NULL, plot = !is.null(coord.cols) || inherits(data, "SpatVector"),
df = TRUE, verbosity = 2)
}
\arguments{
  \item{data}{a 'data.frame' or an object that can be coerced to such (e.g. a 'matrix', 'tibble', 'SpatVector' or 'sf' object) containing a column with the species' presence (1) and absence (0) records, with localities in rows; and (if distance or spatial bias are required) two columns with the spatial coordinates, x and y.}
  \item{sp.cols}{names or index numbers of the columns containing the species presences (1) and absences (0) in 'data'.}
  \item{coord.cols}{names or index numbers of the columns containing the spatial coordinates in 'data' (x and y, or longitude and latitude, in this order!). Needed if distance or spatial bias are required.}
\item{CRS}{coordinate reference system of the 'coord.cols' in 'data', in one of the following formats: WKT/WKT2, <authority>:<code>, or PROJ-string notation (see \code{\link[terra:crs]{terra::crs()}}). Ignored if 'dist.mat' is provided. Otherwise, if 'CRS' is provided and the 'terra' package is installed, distances are computed with \code{\link[terra:distance]{terra::distance()}}, thus accounting for the curvature of the Earth.}
  \item{min.dist}{(optional) numeric value specifying the minimum distance (in the same units as 'coord.cols') at which selected absences should be from the presences.}
  \item{max.dist}{(optional) numeric value specifying the maximum distance (in the same units as 'coord.cols') at which selected absences should be from the presences.}
  \item{n}{(optional) integer value specifying the number of absence rows to select. Can also be specified as a ratio -- see 'mult.p' below.}
  \item{mult.p}{(optional) numeric value specifying how many times the number of presences to use as 'n' (e.g. 10 times as many absences as presences). Ignored if 'n' is not NULL.}
  \item{bias}{either a logical value TRUE to make the selection of absences biased towards the vicinity of presences (which requires specifying 'coord.cols'; may take time and memory for large datasets if 'dist.mat' is not provided); or a SpatRaster layer (quantifying e.g. survey effort, accessibility or human presence; see Details and \code{\link{biasLayer}}) with higher pixel values where the selection of absences should be more likely. Note that this layer should have (approximately) the same spatial resolution as 'data'. The default is FALSE, for no bias.}
  \item{dist.mat}{(if \code{bias=TRUE}) optional argument to pass to \code{\link{distPres}}. If not provided, a distance matrix will be computed with \code{\link[stats:dist]{stats::dist()}}) or with \code{\link[terra:distance]{terra::distance()}}).}
  \item{bunch}{[PENDING IMPLEMENTATION] logical value specifying if the selected absences should concentrate around presences in proportion to their local density, as in Vollering et al. (2019). The default is FALSE.}
  \item{seed}{(optional) integer value to pass to \code{\link{set.seed}} specifying the random seed to use for sampling among the input absences.}
  \item{plot}{logical value specifying whether to plot the result. The default is TRUE if 'data' is a SpatVector or if 'coord.cols' are provided, and FALSE otherwise.}
  \item{df}{logical value specifying whether to return a dataframe with the input 'data' after removal of the non-selected absences. The default is TRUE. If set to FALSE, the output is a logical vector specifying if each row of 'data' was selected or not.}
  \item{verbosity}{numeric value (default 2, for maximum) indicating the amount of messages to display. Choose 0 for no messages.}
}
\details{
Species occurrence data typically incorporate two probability distributions: the actual probability of the species being present, and the probability of it being recorded if it was present (Merow et al. 2013). Thus, any covariation between recording probability and the predictor variables can bias the predictions of species distribution models (Yackulic et al. 2013).

Methods to correct for this bias include the selection of (pseudo)absences preferably towards the vicinity of presence records, in order to reproduce the survey bias. This function implements this strategy in several (alternative or complementary) ways: 1) selecting absences within and/or beyond a given distance from presences; 2) biasing the random selection of absences, making it more likely towards the vicinity of presences (providing the 'prob' argument in \code{\link{sample}} with the result of \code{\link{distPres}}); or [PENDING IMPLEMENTATION!] 3) bunching up the absences preferably around the areas with higher densities of presences (Vollering et al. 2019).

More recent versions of this function allow using instead a raster map of weights (bias layer), with higher pixel values where absence selection should be proportionaly more likely, and zero or NA where pseudoabsence points should not be placed. This layer should normally reflect (a proxy for) likely survey effort, such as proximity to roads or populated areas, human footprint (see e.g. \code{\link[geodata:footprint]{geodata::footprint()}}), travel time (e.g. \code{\link[geodata:travel_time]{geodata::travel_time()}}), or the number of species or records of the same taxonomic group (El-Gabbas, 2026). You can also create your own e.g. with \code{\link{biasLayer}}. Users should provide the bias layer at aproximately the same spatial resolution as the 'coord.cols' in the input 'data'.
}
\value{
This function returns the 'data' input after removal of the non-selected absences, or (if df=FALSE) a logical vector specifying if each row of 'data' was selected (TRUE) or not (FALSE).

If plot=TRUE and provided 'coord.cols' or SpatVector 'data', the function also plots the presences (blue "plus" signs), the selected absences (red "minus" signs) and the excluded absences (orange dots).
}
\references{
El-Gabbas A. (2026) A Global, Taxon-Stratified, High-Resolution Sampling-Effort Dataset From GBIF for Bias-Aware Ecological Modelling. Diversity and Distributions 32(5): e70205

Vollering J., Halvorsen R., Auestad I. & Rydgren K. (2019) Bunching up the background betters bias in species distribution models. Ecography, 42: 1717-1727
}
\author{
A. Marcia Barbosa
}
\seealso{
\code{\link{gridRecords}}, \code{\link{sample}}, \code{\link{biasLayer}}
}
\examples{
data(rotif.env)

head(rotif.env)

names(rotif.env)

table(rotif.env$Burceo)

# select among the absences using different criteria:

abselect <- selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), n = 150, seed = 123)

abselect <- selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), mult.p = 1.5, seed = 123)

\dontrun{
data(rotif.env)

abselect <- selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), max.dist = 18)

abselect <- selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), max.dist = 18, min.dist = 5,
n = sum(rotif.env$Burceo), bias = TRUE)


# you can run the next example if you have the 'geodata' package installed:

human_footprint <- geodata::footprint(year = 2009, path = tempfile())

# see also the biasLayer() function for creating your own bias layer
# or see below for creating a bias layer based on target-group background


abselect <- fuzzySim::selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), n = sum(rotif.env$Burceo),
bias = human_footprint)

terra::plot(human_footprint, alpha = 0.3, add = TRUE)


# you can run the next example if you have 'terra' and 'spatstat' installed:

# coordinates of records (surveyed sites) for the whole taxonomic group:
recs <- rotif.env[rowSums(rotif.env[ , 18:47] > 0), c("Longitude", "Latitude")]

rng <- terra::ext()

win <- spatstat.geom::owin(xrange = rng[1:2], yrange = rng[3:4])

ppp <- spatstat.geom::ppp(x = recs$Longitude, y = recs$Latitude, window = win)

kernel_density <- terra::rast(spatstat.explore::density.ppp(ppp))

terra::plot(kernel_density)

points(recs, col = "red", cex = 0.2)

abselect <- fuzzySim::selectAbsences(data = rotif.env, sp.cols = "Burceo",
coord.cols = c("Longitude", "Latitude"), n = sum(rotif.env$Burceo),
bias = kernel_density)

terra::plot(kernel_density, alpha = 0.3, add = TRUE)


# try also the sampling effort layers published in:
# El-Gabbas A. 2026. A Global, Taxon-Stratified, High-Resolution Sampling-Effort
# Dataset From GBIF for Bias-Aware Ecological Modelling. Diversity and
# Distributions 32(5): e70205. https://doi.org/10.1111/ddi.70205.
}
}
